requirements.txt
fastapi==0.87.0
django-environ==0.10.0
uvicorn==0.20.0
requests==2.28.1
